requirements.txt
numpy
boto3-stubs[s3,opensearch]
farm-haystack[aws,preprocessing,opensearch,inference]
sagemaker
